added player sprite, fixed astar, and spritesheet

+++ LostInChaos/src/Bullet.cpp
@@ -9,8 +9,8 @@ Bullet::Bullet(float x, float y, Object* target, SDL_Renderer* renderer, std::st
 	int xp, yp;
 	SDL_GetMouseState(&xp, &yp);
 
-	double diffX = x - xp;
-	double diffY = y - yp;
+	double diffX = x - (double) xp;
+	double diffY = y - (double) yp;
 
 	double* unitVectors = calcVector(diffX, diffY);
 	unitX = unitVectors[0];
@@ -25,15 +25,15 @@ void Bullet::move(std::array<Tile*, MAP_LENGTH> &map, double deltaTime) {
 	}
 	else {
 		// if bullet is in screen , move it
-		if (y > 0) y -= unitY * (deltaTime / 7);
+		if (y > 0) y -= (float) (unitY * (deltaTime / 7));
 		if (y > SCREEN_HEIGHT) alive = false;
 
-		if (x > 0 && x < SCREEN_WIDTH - collisionRect.w) x -= unitX * (deltaTime / 7);
+		if (x > 0 && x < SCREEN_WIDTH - collisionRect.w) x -= (float) (unitX * (deltaTime / 7));
 		else alive = false;
 	}
 
-	collisionRect.x = x;
-	collisionRect.y = y;
+	collisionRect.x = (int) x;
+	collisionRect.y = (int) y;
 }
 
 void Bullet::render() {
